backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy>=2.0.35  # Updated for Python 3.13 compatibility
psycopg2-binary>=2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
# hiredis==2.2.3  # Optional: requires C++ build tools. Install separately if needed.

# Async HTTP
httpx==0.25.2
# aiohttp==3.9.1  # Optional: requires C++ build tools. httpx is sufficient.

# Background tasks
celery==5.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# LLM SDKs
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.1
cohere==4.37

# Vector DB (RAG System) - Optional: requires C++ build tools
# chromadb==0.4.22  # Install separately: pip install chromadb (requires C++ build tools)
# sentence-transformers==2.2.2  # Install separately: pip install sentence-transformers (requires C++ build tools)
# Alternative: Use simpler vector DB or install build tools first

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

